Cleaned up the code for the Timer.
